support negative horizontal widths in man(7); minus twenty lines of code in spite of enhanced functionality

+++ term.h
@@ -121,10 +121,8 @@ void		  term_begin(struct termp *, term_margin,
 void		  term_end(struct termp *);
 
 void		  term_setwidth(struct termp *, const char *);
-size_t		  term_hspan(const struct termp *,
-			const struct roffsu *);
-size_t		  term_vspan(const struct termp *,
-			const struct roffsu *);
+int		  term_hspan(const struct termp *, const struct roffsu *);
+int		  term_vspan(const struct termp *, const struct roffsu *);
 size_t		  term_strlen(const struct termp *, const char *);
 size_t		  term_len(const struct termp *, size_t);
